feat: theme에 PriceChart 관련 정보 등록
